Use specialized GPIO_PIN for leds in the blink example
Added a specialized GPIO_PIN_LED which allows for an LED to be active high
or active low.  Additional functions on() and off() utilize the active
flag to set the pin to either high or low.

Also, added the ability to set an LED=1; or LED=0 with an operator overload.
Using 1 sets it to on() and 0 to off() or you can use any of the enumerators
HIGH, LOW, ON, OFF